Add index and first recipe html files
I have populated the main homepage for the project that will eventually link to multiple different cooking recipes. Currently, it only links to a recipe for beans, which is in a parent directory recipes/. Subsequent recipe html files should go in this directory and will be added to the unordered list.